Written some more views. Auction should be functionally complete. Orders can now be placed

+++ src/main/java/com/ingsw/flyingdutchman/controller/ProductManagement.java
@@ -186,6 +186,7 @@ public class ProductManagement {
                         Integer.parseInt(request.getParameter("starting_price")),
                         Integer.parseInt(request.getParameter("current_price")),
                         null, // TODO: se ho tempo implementare lettura immagini
+                        false,
                         categoryDAO.findByCategoryID(Long.parseLong(request.getParameter("categoryID"))),
                         user
                 );
@@ -285,9 +286,25 @@ public class ProductManagement {
 
             ProductDAO productDAO = daoFactory.getProductDAO();
             Product product1 = productDAO.findByProductID(Long.parseLong(request.getParameter("productID")));
-            productDAO.delete(product1);
 
             loggedUser = daoFactory.getUserDAO().findByUsername(loggedUser.getUsername());
+
+            //Cerco se il prodotto è attualmente in vendita all'asta
+            //---
+            //Cerco tutte le aste con quel prodotto all'interno
+            Auction[] auctionsToCheck = daoFactory.getAuctionDAO().findByProductOwner(product1);
+
+            //Vedo se in quelle aste è presente quel prodotto ancora invenduto
+            boolean isCurrentlyAuctioned = false;
+            for(int i = 0; i < auctionsToCheck.length ; i++){
+                if(auctionsToCheck[i].getClosing_timestamp() == null){
+                    applicationMessage = "Non puoi eliminare un prodotto attualmente all'asta!";
+                    isCurrentlyAuctioned = true;
+                }
+            }
+            if(!isCurrentlyAuctioned){
+                productDAO.delete(product1);
+            }
 
             // Prendo tutte le aste con prodotti dell'utente e le riempio
             Auction[] auctions = daoFactory.getAuctionDAO().findByOwner(loggedUser);
